requirements.txt
Django>=1.10
sqlparse
python_dateutil
gunicorn
whitenoise
celery>=3.1,<4.0
django-celery
PyYAML
pyzmq
msgpack-python
tabulate
